pyproject.toml
[project]
name = "msc-mlai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clip>=0.2.0",
    "gensim>=4.3.3",
    "hydra-core>=1.3.2",
    "lightning>=2.5.5",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "nltk>=3.9.1",
    "open-clip-torch>=3.2.0",
    "pandas>=2.3.2",
    "pytorch-lightning>=2.5.4",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.13.1",
    "shortuuid>=1.0.13",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "wandb>=0.21.3",
]
